requirements.txt
amqp==5.1.1
asgiref==3.5.0
async-timeout==4.0.2
autopep8==1.6.0
backports.zoneinfo==0.2.1
backtrader==1.9.76.123
backtrader-plotly==1.2.0
billiard==3.6.4.0
celery==5.2.6
celery-progress==0.1.2
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==4.0.4
django-debug-toolbar==3.2.4
djangorestframework==3.13.1
djangorestframework-datatables==0.7.0
idna==3.3
kombu==5.2.4
numpy==1.22.3
packaging==21.3
pandas==1.4.2
plotly==5.7.0
prompt-toolkit==3.0.29
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2022.1
redis==4.2.2
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tenacity==8.0.1
toml==0.10.2
urllib3==1.26.9
uWSGI==2.0.20
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.0
